tacotron2_gst/requirements.txt
fire>=0.3
numpy>=1.13
torch>=1.3
matplotlib>=2.1
inflect>=0.2
librosa==0.6.0
unidecode>=1.1
pillow>=7.1
pyyaml>=5.3
scikit-learn>=0.23.1
numba==0.48
tensorboard==1.15.0
tqdm>=4.61.1
